requirements.txt
matplotlib
pyserial
datetime